requirements.txt
Django==3.0
channels==2.4
channels_redis
Pillow==7.0.0
sphinx
djangorestframework
djangorestframework-simplejwt

